Refactor: Resource -> Asset으로 수정 (#27)
