+++ getAsset/getEC2.py
